refactor: VenueMapper dummy 사용 코드 제거

+++ be/src/main/java/kr/codesquad/jazzmeet/venue/dto/response/VenueSearchResponse.java
@@ -10,7 +10,7 @@ public record VenueSearchResponse(
 	List<VenueSearch> venues,
 	long totalCount,
 	int currentPage,
-	long maxPage) {
+	int maxPage) {
 	public static VenueSearchResponse emptyVenues() {
 		return VenueSearchResponse.builder().venues(List.of()).build();
 	}
